Spaces: add oprhan option to destory to leave control planes

+++ cmd/up/space/destroy.go
@@ -30,7 +30,7 @@ import (
 )
 
 const (
-	confirmStr      = "DESTROY CONFIRMED"
+	confirmStr      = "CONFIRMED"
 	nsUpboundSystem = "upbound-system"
 )
 
@@ -41,15 +41,25 @@ func (c *destroyCmd) AfterApply(insCtx *install.Context) error {
 	upterm.DefaultObjPrinter.Pretty = true
 
 	if !c.Confirmed {
+		if c.Orphan {
+			pterm.Info.Println()
+			pterm.Info.Println("Removing Space API components.")
+			pterm.Info.Println("Control Planes will continue to run and no data will be lost")
+			pterm.Info.Println()
+		} else {
+			pterm.Println()
+			pterm.FgRed.Println("******************** DESTRUCTIVE COMMAND ********************")
+			pterm.FgRed.Println("********************* DATA-LOSS WARNING *********************")
+			pterm.Println()
+			pterm.Warning.Println("Destroying Spaces is a destructive command that will destroy data and oprhan resources.")
+			pterm.Warning.Println("Before proceeding ensure that Managed Resources in Control Planes have been deleted.")
+			pterm.Warning.Println("All Spaces components including Control Planes will be destroyed.")
+			pterm.Println()
+			pterm.Warning.Println("If you want to retain data, abort and run 'up space destroy --orphan'")
+			pterm.Println()
+		}
+
 		prompter := input.NewPrompter()
-		pterm.Println()
-		pterm.FgRed.Println("******************** DESTRUCTIVE COMMAND ********************")
-		pterm.FgRed.Println("****************** DATA-DESTRUCTION WARNING *****************")
-		pterm.Println()
-		pterm.Warning.Println("Destroying Spaces is a destructive command that will destroy data and oprhan resources.")
-		pterm.Warning.Println("Before proceeding ensure that Managed Resources in Control Planes have been deleted.")
-		pterm.Warning.Println("All Spaces components including Control Planes will be destroyed.")
-		pterm.Println()
 		in, err := prompter.Prompt(fmt.Sprintf("To proceed, type: %q", confirmStr), false)
 		if err != nil {
 			pterm.Error.Printfln("error getting user confirmation: %v", err)
@@ -67,14 +77,23 @@ func (c *destroyCmd) AfterApply(insCtx *install.Context) error {
 	}
 	c.kClient = kClient
 
+	with := []helm.InstallerModifierFn{
+		helm.WithNamespace(ns),
+		helm.IsOCI(),
+	}
+	if c.Orphan {
+		with = append(with, helm.WithNoHooks())
+	}
+
 	mgr, err := helm.NewManager(insCtx.Kubeconfig,
 		spacesChart,
 		c.Registry,
-		helm.WithNamespace(ns),
-		helm.IsOCI())
+		with...,
+	)
 	if err != nil {
 		return err
 	}
+
 	c.mgr = mgr
 	return nil
 }
@@ -86,13 +105,19 @@ type destroyCmd struct {
 
 	commonParams
 
-	Confirmed bool `name:"yes-really-delete-spaces-and-all-data" type:"bool" help:"Bypass safety checks and destroy Spaces"`
+	Confirmed bool `name:"yes-really-delete-space-and-all-data" type:"bool" help:"Bypass safety checks and destroy Spaces"`
+	Orphan    bool `name:"orphan" type:"bool" help:"Remove Space components but retain Control Planes and data"`
 }
 
 // Run executes the uninstall command.
 func (c *destroyCmd) Run(insCtx *install.Context) error {
 	if err := c.mgr.Uninstall(); err != nil {
 		return err
+	}
+	// leave `upbound-system` namespace in place since there are secrets, configmaps, etc,
+	// used by controlplanes
+	if c.Orphan {
+		return nil
 	}
 	return c.kClient.CoreV1().Namespaces().Delete(context.Background(), nsUpboundSystem, v1.DeleteOptions{})
 }
